Like it it not, it's maven time.
If we're going to work on this code, we're going to need a build system,
so I've crowbared maven into this repo :) Some other small changes
dotted around too.

Changelog;

REFACTOR: "elucidusFilterPdf" -> "elucidusGeneratorPdf" - to match the
new naming convention

NEW: .poms for the projects, including dependancies.

REFACTOR: Empty JUnit tests are now ignored, not failed - this allows
for a clean maven build from top to bottom.
-