Added the first spooler for directories.
Spoolers sit like a daemon and "listen" for files to feed in to
Elucidus. This spooler just watches a local directory (taken as a first
argument) and routes files to generators based on the their file
extension.

Lots of work and improvements possible, but I'm just stubbing a bunch of
stuff at the moment - want to get an end to end testable system
available.

Some "API" changes were necessary;

Item: Added a pretty toString. Pretty sure the PDF generator does
something similar but I could not be arsed to look where.

(new class) StandardAspects. Provides Java constants for frequently
used/standard elucidus aspects. Made this with a bunch of psf strings,
but maybe an ENUM would be more inteligent for the future - you could do
some cool stuff (tm).

Hacked the GeneratorFactory to within an inch of it's life, sorry about
that, as I said, stubbing in a bunch of stuff at the moment, it didn't
put up much of a fight!